+++ nim_core/ios/Classes/FLTService.swift
@@ -22,6 +22,12 @@ enum ServiceType: String {
   case SettingService = "SettingsService"
   case SuperTeamService
   case AvSignallingService
+  case QChatObserver
+  case QChatService
+  case QChatServerService
+  case QChatChannelService
+  case QChatMessageService
+  case QChatRoleService
 }
 
 protocol FLTService {