+++ nim_core_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.3.0(Nov 3, 2022)
+### New Features
+* Support IM QChat in iOS and Android
+
+## 1.0.4(Sep 30, 2022)
+### Bug Fixes
+* Fix enum .name error in sdk less than 2.15.0
+
+## 1.0.3(Sep 30, 2022)
+### New Features
+* Support IM Signalling in iOS and Android
+
 ## 1.0.2(Sep 23, 2022)
 ### New Features
 * Support Web Plugin